🚀 feat(deployCollectionWrapper.ts): add deployCollectionWrapper function to handle the deployment of NFT collections 🚚 chore(EventPassDeployButtonClient.tsx): move deployCollectionWrapper to actions folder for better organization 🚚 chore(EventSheet.stories.tsx): move deployCollectionWrapper and renameEventPassNftFiles to actions folder for better organization 🔧 chore(EventSheet.stories.tsx): mock renameEventPassNftFiles and deployCollectionWrapper functions for testing purposes 🔧 chore(EventSheet.stories.tsx): add organizerId prop to args in EventSheetExample for testing purposes 🔧 chore(EventSheet.stories.tsx): mock getSigner and provider properties in useAuthContext for testing purposes 🔧 chore(EventSheet.stories.tsx): add test case for clicking on deploy button and assert that deployCollectionWrapper is called with the correct arguments
